Move schema decorators classes out of closures
Closes: https://github.com/metarhia/metaschema/issues/47
