+++ lib/schema.js
@@ -6,6 +6,7 @@ const { basename } = require('path');
 const metasync = require('metasync');
 
 const { parseSignature } = require('./introspection.js');
+const decorators = require('./decorators.js');
 
 const SCRIPT_TIMEOUT = 2000;
 
@@ -222,92 +223,18 @@ const enumCollection = values => {
   };
 };
 
-const Enum = (...args) => {
-  const item = args[0];
-  const itemType = typeof(item);
-  if (itemType === 'object') return enumCollection(item);
-  if (itemType !== 'string') return enumArray(args);
-  return enumCollection(Object.assign({}, args));
-};
-
-const List = category => items => items.map(
-  item => createInstance(category, item)
-);
-
-const Many = def => {
-  class Many {
-    constructor(def) {
-      Object.assign(this, def);
-    }
-  }
-  return new Many(def);
-};
-
-const Parent = def => {
-  class Parent {
-    constructor(def) {
-      Object.assign(this, def);
-    }
-  }
-  return new Parent(def);
-};
-
-const Master = def => {
-  class Master {
-    constructor(def) {
-      Object.assign(this, def);
-    }
-  }
-  return new Master(def);
-};
-
-const Include = def => {
-  class Include {
-    constructor(def) {
-      Object.assign(this, def);
-    }
-  }
-  return new Include(def);
-};
-
-const Dictionary = def => {
-  class Dictionary {
-    constructor(def) {
-      Object.assign(this, def);
-    }
-  }
-  return new Dictionary(def);
-};
-
-const Index = (...fields) => {
-  class Index {
-    constructor(fields) {
-      this.fields = fields;
-    }
-  }
-  return new Index(fields);
-};
-
-const Unique = (...fields) => {
-  class Unique {
-    constructor(fields) {
-      this.fields = fields;
-    }
-  }
-  return new Unique(fields);
-};
-
-const Validate = fn => {
-  class Validate {}
-  return Object.setPrototypeOf(fn, Validate);
-};
-
-Object.assign(structs, {
-  Enum, List, // structure decorators
-  Many, Parent, Master, Include, // field decorators
-  Dictionary, // category decorators
-  Index, Unique, // index decorators
-  Validate // validation function decorator
+Object.assign(structs, decorators, {
+  // structure decorators
+  Enum: (...args) => {
+    const item = args[0];
+    const itemType = typeof(item);
+    if (itemType === 'object') return enumCollection(item);
+    if (itemType !== 'string') return enumArray(args);
+    return enumCollection(Object.assign({}, args));
+  },
+  List: category => items => items.map(
+    item => createInstance(category, item)
+  ),
 });
 
 module.exports = {
